requirements.txt
streamlit
pandas
plotly
google-generativeai
python-dotenv
fpdf2
openpyxl
numpy